requirements.txt
django==2.0.7
django-debug-toolbar==1.9.1
django-bootstrap4==0.0.6